feat: Android v1.1.0 release build - fix compilation errors and configure release
- Bump version to 1.1.0 (versionCode 2)
- Fix 30+ Kotlin compilation errors (smart cast, imports, API changes, type mismatches)
- Configure Alibaba Cloud Maven mirrors for dependency resolution
- Add release signing keystore and V2 signing
- Add kotlin-parcelize plugin for @Parcelize support
- Fix AndroidManifest shortcut definition (move to res/xml/shortcuts.xml)
- Fix VectorDrawable circle element (replace with path arc)
- Configure lint and ProGuard for release build compatibility
- Fix Camera2 API constant references and MediaCodecInfo API changes

+++ LiveCaptureAndroid/app/src/main/java/com/livecompose/livecapture/core/frame/WatermarkInfo.kt
@@ -12,7 +12,7 @@ data class WatermarkInfo(
 
     // 文字样式
     val textSize: Float = 14f,
-    val textColor: Int = 0x80FFFFFF,     // 半透明白色
+    val textColor: Int = 0x80FFFFFF.toInt(),     // 半透明白色
     val fontName: String = "default",     // default / ds-digital (数码字体)
 
     // 位置
